requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
opensearch-py
psycopg2-binary
pydantic
python-dotenv
pypdf
tqdm